backend/requirements.txt
# Core AI & Agent Frameworks
google-adk>=0.1.0           # The Agent Development Kit
google-genai>=0.3.0         # Gemini 2.0 SDK
python-a2a>=0.1.0           # Agent-to-Agent Protocol
mcp>=0.1.0                  # Model Context Protocol

# Web Framework (API)
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Data & Config
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Observability & Testing
opentelemetry-sdk>=1.20.0
pytest>=7.0.0
nest_asyncio>=1.5.0         # For Jupyter/Loop handling